WIP adding more flexible dtype enforcement
