+++ tests/utils/test_misc.py
@@ -98,3 +98,27 @@ def test_filter_unseen_df(old_docs, new_docs, expected_docs):
     new_df = pd.DataFrame(new_docs)
     delta_df = filter_unseen_df(old_df, new_df)
     assert delta_df.to_dict(orient='records') == expected_docs
+
+
+@pytest.mark.parametrize(
+    'ldtype,rdtype,are_equal',
+    [
+        ('string', 'string', True),
+        ('str', 'string', True),
+        ('str', 'object', True),
+        ('int', 'int32', True),
+        ('datetime64[ns, UTC]', 'Timestamp', True),
+        ('float', 'float64', True),
+        ('bool', 'bool[pyarrow]', True),
+        ('Int64', 'int', True),
+        ('json', 'object', True),
+        ('float', 'object', False),
+        ('datetime', 'object', False),
+    ]
+)
+def test_are_dtypes_equal(ldtype: str, rdtype: str, are_equal: bool):
+    """
+    Test that different combinations of dtypes are equal (or inequal).
+    """
+    from meerschaum.utils.misc import are_dtypes_equal
+    assert are_dtypes_equal(ldtype, rdtype) == are_equal
